Feature/erecord/15 add discount functionality to basket controller (#16)
* Changed Discount.Percentage to int

* Updated command tests with expected values

* Registered IBasketDiscountService

* Added BasketDiscountService and QueryTotalCostOfBasketCommand to basketControllerFacade

* Changed product price to 2 decimal places

* Added SetDiscountCode and GetBasketTotalCost routes

* Added TotalCostAfterDiscount route

* Updated interface method name

* Updated tests

* Fixed bug where updating basket productIds would reset discount code

* Updated visibility

* Updated variable name

* Updated dev-logs and notes

* Updated cSpell words

* Removed extra space

+++ backend/tests/Factories/TestDiscountFactory.cs
@@ -5,7 +5,7 @@ namespace StoreBackend.Tests.Factories
 {
     public static class TestDiscountFactory
     {
-        public static Discount CreateDiscount(int discountId, string discountCode, decimal discountPercentage) => new Discount
+        public static Discount CreateDiscount(int discountId, string discountCode, int discountPercentage) => new Discount
         {
             Id = discountId,
             Code = discountCode,
